New layout and added "Contact the Team"-button
Moved dog logo from under the "Fork me on GitHub"-button in order to
click it, changed font and added popup to logo when hovered over. Also
added orange border to map and made the map a bit more narrow. Added
three new buttons (or at least a new look); "+ Add new Marker" (locating
function is still in the code, but instead of locating and placing the
marker, it gives an alert that the function is not fully implemented
yet), "Contact the Team" and "@Pick-a-Poop on Twitter". They're all
connected and in the same row now.